+++ src/shared.py
@@ -8,11 +8,11 @@ from pygame.locals import *
 tileSize               = 32
 
 heroWalkingSpriteTempo = 4
-heroWalkingSpeed       = tileSize / 8
+heroWalkingSpeed       = int(tileSize / 8)
 heroAttackSpriteTempo  = 2
 
 ennemyWalkingSpeed     = 1
-ennemyKnockBack        = tileSize / 3
+ennemyKnockBack        = int(tileSize / 3)
 
 debug = False
 #debug = True
